Move getTextContent from urbit/ -> client/
